tool_library/example_tools/chat_ui/requirements.txt
openai
chainlit==1.0.200
fastapi
uvicorn
langchain
anvil-app-server
anvil-uplink